Implement PaperRepositoryList()
